Mon, Jul 31, 2023, 10:58 AM +04:30

+++ CS50x/PSET2/readability/readability.c
@@ -33,7 +33,7 @@ int main(void)
     }
     int L = numC / numW * 100;
     int S = numS / numW * 100;
-    int index = 0.0588 * L - 0.296 * S - 15.8;
+    float index = 0.0588 * L - 0.296 * S - 15.8;
     if (index >= 16)
     {
         printf("Grade 16+\n");
@@ -44,7 +44,7 @@ int main(void)
     }
     else
     {
-        printf("Grade %i\n", index);
+        printf("Grade %f\n", index);
     }
     printf("%i %i %i\n", numC, numW, numS);
